make add role work again
